Prepare for new version.
Please Stand by...

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/windows/WndWandmaker.java
@@ -3,7 +3,7 @@
  * Copyright (C) 2012-2015 Oleg Dolya
  *
  * Shattered Pixel Dungeon
- * Copyright (C) 2014-2019 Evan Debenham
+ * Copyright (C) 2014-2021 Evan Debenham
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -22,12 +22,11 @@
 package com.shatteredpixel.shatteredpixeldungeon.windows;
 
 import com.shatteredpixel.shatteredpixeldungeon.Dungeon;
-import com.shatteredpixel.shatteredpixeldungeon.actors.hero.HeroClass;
 import com.shatteredpixel.shatteredpixeldungeon.actors.mobs.npcs.Wandmaker;
 import com.shatteredpixel.shatteredpixeldungeon.items.Item;
-import com.shatteredpixel.shatteredpixeldungeon.items.potions.PotionOfExperience;
 import com.shatteredpixel.shatteredpixeldungeon.items.quest.CorpseDust;
 import com.shatteredpixel.shatteredpixeldungeon.items.quest.Embers;
+import com.shatteredpixel.shatteredpixeldungeon.items.wands.Wand;
 import com.shatteredpixel.shatteredpixeldungeon.messages.Messages;
 import com.shatteredpixel.shatteredpixeldungeon.plants.Rotberry;
 import com.shatteredpixel.shatteredpixeldungeon.scenes.PixelScene;
@@ -44,66 +43,68 @@ public class WndWandmaker extends Window {
 	private static final float GAP		= 2;
 	
 	public WndWandmaker( final Wandmaker wandmaker, final Item item ) {
-
+		
 		super();
-
+		
 		IconTitle titlebar = new IconTitle();
 		titlebar.icon(new ItemSprite(item.image(), null));
 		titlebar.label(Messages.titleCase(item.name()));
 		titlebar.setRect(0, 0, WIDTH, 0);
-		add(titlebar);
+		add( titlebar );
 
 		String msg = "";
-
-		if (item instanceof CorpseDust) {
+		if (item instanceof CorpseDust){
 			msg = Messages.get(this, "dust");
-		} else if (item instanceof Embers) {
+		} else if (item instanceof Embers){
 			msg = Messages.get(this, "ember");
-		} else if (item instanceof Rotberry.Seed) {
+		} else if (item instanceof Rotberry.Seed){
 			msg = Messages.get(this, "berry");
 		}
 
-		RenderedTextBlock message = PixelScene.renderTextBlock(msg, 6);
+		RenderedTextBlock message = PixelScene.renderTextBlock( msg, 6 );
 		message.maxWidth(WIDTH);
 		message.setPos(0, titlebar.bottom() + GAP);
-		add(message);
-
-
-		RedButton btnWand1 = new RedButton(Messages.titleCase(Wandmaker.Quest.wand1.name())) {
+		add( message );
+		
+		RedButton btnWand1 = new RedButton( Messages.titleCase(Wandmaker.Quest.wand1.name()) ) {
 			@Override
 			protected void onClick() {
-				selectReward(wandmaker, item, Wandmaker.Quest.wand1);
+				selectReward( wandmaker, item, Wandmaker.Quest.wand1 );
 			}
 		};
 		btnWand1.setRect(0, message.top() + message.height() + GAP, WIDTH, BTN_HEIGHT);
-		add(btnWand1);
-
-		RedButton btnWand2 = new RedButton(Messages.titleCase(Wandmaker.Quest.wand2.name())) {
+		add( btnWand1 );
+		
+		RedButton btnWand2 = new RedButton( Messages.titleCase(Wandmaker.Quest.wand2.name()) ) {
 			@Override
 			protected void onClick() {
-				selectReward(wandmaker, item, Wandmaker.Quest.wand2);
+				selectReward( wandmaker, item, Wandmaker.Quest.wand2 );
 			}
 		};
 		btnWand2.setRect(0, btnWand1.bottom() + GAP, WIDTH, BTN_HEIGHT);
-		add(btnWand2);
-
+		add( btnWand2 );
+		
 		resize(WIDTH, (int) btnWand2.bottom());
 	}
 	
-	private void selectReward( Wandmaker wandmaker, Item item, Item reward ) {
-		
+	private void selectReward( Wandmaker wandmaker, Item item, Wand reward ) {
+
+		if (reward == null){
+			return;
+		}
+
 		hide();
-		
+
 		item.detach( Dungeon.hero.belongings.backpack );
 
 		reward.identify();
-		if (reward.doPickUp(Dungeon.hero)) {
-			GLog.i(Messages.get(Dungeon.hero, "you_now_have", reward.name()));
+		if (reward.doPickUp( Dungeon.hero )) {
+			GLog.i( Messages.get(Dungeon.hero, "you_now_have", reward.name()) );
 		} else {
-			Dungeon.level.drop(reward, wandmaker.pos).sprite.drop();
+			Dungeon.level.drop( reward, wandmaker.pos ).sprite.drop();
 		}
 		
-		wandmaker.yell( Messages.get(this, "farewell", Dungeon.hero.givenName()) );
+		wandmaker.yell( Messages.get(this, "farewell", Dungeon.hero.name()) );
 		wandmaker.destroy();
 		
 		wandmaker.sprite.die();
